requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
magic-filter==1.0.9
multidict==6.0.4
peewee==3.15.4
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
yarl==1.8.2
